CDP 1101: Add set_harvest_fee function (#94)
* CDP 1101: Add set_harvest_fee function

* CDP 1101: Add set_harvest_fee function

* touch ups

* formatting and adding comments wherever necessary

+++ programs/stable-pool/src/states/global_state.rs
@@ -11,9 +11,11 @@ pub struct GlobalState {
     pub mint_usdx: Pubkey,
     /// Bump/nonce for the USDx mint address
     pub mint_usdx_bump: u8,
+    /// TODO: rename to tvl_limit_usd
+    /// aliases: tvlLimit
     pub tvl_limit: u64,
     /// The total value locked on the CDP platform, in USD
-    /// 
+    ///
     /// TODO: determine how to calculate tvl from each type of collateral
     pub tvl_usd: u64,
     /// Is contract paused
@@ -25,11 +27,12 @@ pub struct GlobalState {
     /// The limit on the mintable debt per user, in USDx
     pub user_debt_ceiling: u64,
     /// The numerator for calculating the fee
-    pub fee_num: u32,
+    /// TODO: rename to harvest_fee_numer
+    pub fee_num: u64,
     /// The denomenator for calculating the fee
-    pub fee_deno: u32,
+    pub fee_deno: u64,
     /// The collateral per risk
-    pub coll_per_risklv: [u64; 10],
+    pub coll_per_risklv: [u64; 10], // can we rename this
 
     /// only this wallet can report new prices to the oracle accounts
     pub oracle_reporter: Pubkey,
